Minor fixes - non messed-up version!
A few undeclared variables - no big deal there. One use of block_end_line where block.end_line was meant (I believe).

Also - query as to whether the new keyword is needed when calling InlineParser as the function returns an object, so we're not dealing with an instance of InlineParser (I haven't changed this, just a query).