Changed short name from `stmd` to `cmark`.
Summary:

PR #87 does this as part of a much larger set of changes, but I thought
it was important enough to move to a bite-sized commit in order to push
it through.

Though this touches many files, any conflicts should be very
straightforward to fix.

Test Plan:

Ran `make all` and then `make test && make testjs` and observed all
green.