Removed self links
The spec has a bunch of links that redirect to themselves. This is
confusing, and it's not much of a link. I assume this is done to point
out that this is where the term is being defined. I modified those
occurrences to be bolded instead. Moreover, they were usually followed
by "id links". If those were referenced elsewhere, a user would be
redirected to the middle of a section. Being redirected to the beginning
of the section is much clearer. Here are the rules I tried to follow:

- Removed all links that redirected to the same section the link was in.
- When the linked text was the term being defined in the section, bolded
it.
- Moved all "id links" to right before the section header they belong
in.
- If a section is very large, I left the id links next to the term
definitions.

Note that the technique of adding <a id="blabla"></a> before every
header is damned ugly. The spec.html maker adds those id's to the
headers directly, and that seems like it should be default for
CommonMark as a whole. Maybe for future revisions? How about a new {ToC}
element that benefits from that?