Spec:  better system for designating definitions of terms.
`[foo](@foo)` turns into an anchor with id `foo` that links to
itself and has class `definition`.  This allows us to remove the
manual `<a>` tags in `spec.txt`.

A simple regex substitution in the Makefile handles the transformation.

I believe this addresses some of the concerns in #141 by giving
the definitions a prominent appearance.  Note that
we want definitions to link to themselves, to allow users to
quickly generate a link to the relevant bit of the spec.