Restored priority of links over emphasis grouping.
Now when we encounter (possibly) closing `*` or `_` delimiters,
we simply add them to the delimiters stack.

This gets processed by `process_emphasis` either (a) when
a link is created (in which case only the inlines created
by the link are processed) or at the end of processing
a run of inlines.