Clarify that all inputs are valid
This makes it explicitly clear that any input is valid CommonMark and is
expected to be parsed.  The spec didn't previously take an explicit
stance on this, but it can be pieced together by a careful reader.
Reference parsers conform to this.

It's a strong assertion and useful to have in mind when reading about
the details of the format.  I think readers are likely to be used to
more restrictive markup formats (like HTML) which parsers _can_ fail to
parse, so it's good to make this obvious from the start.