Simplified release archive.
We now simply use git archive (which is also what github would
use on their releases page).

This is possible because we now include some generated files in
the repository.

The fact is that this is what people are going to use anyway,
so instead of testing two different setups (our hand-crafted
archive and the git repo), it's better just to use one.