Defer closing of fenced code blocks.
Instead of closing a fenced code block when we see the closing
fence, we set fenceLength to -1, which then signals the code
that checks for block enders that it is time to close the block.

This allows us to process a blank line after the fenced code
and set _lastLineBlank appropriately.

Addresses #285 for JS implementation.