Moved continuation checks & finalizers into blocks property.
This is a first step towards keeping the code for each kind of
block in a central place, rather than spread all over the code
base.

This is preparatory for a more modular structure, where each
type of block has a record describing how it is parsed and
finalized.

Eventually this will also contain functions for checking for
a block start, and metadata that determines how line data
should be handled.

There is a small performance penalty (about 3%?) but it seems
worth it.