Rework Makefile targets
Since $(BUILDDIR) depended on the phony target "check", it was always
considered out-of-date. So it was always rebuilt resulting in running
the "cmake" command again even if it was already run.

Add a new phony target "cmake_build" that always triggers the cmake build
and make $(PROG) depend on it.

Running "make" a second time now doesn't run cmake again.