Preliminary spec changes for new tab handling.
Instead of having a preprocessing step where tabs are
converted to spaces, we now handle tabs directly in
the parser.

This allows tabs to be retained in code blocks and code
spans.

This commit adds some general language to the effect that,
for purposes of determining block structure, tabs are to
be treated just like equivalent spaces.

Much more work is needed on the spec regarding spaces and tabs,
but this change (and corresponding changes to the implementations)
is a start.