Rewrapped two lines so `+` does not begin a line.
Closes #347.

Thanks for pointing this out.

We might want to revisit the decision to allow a list to
interrupt a paragraph.  It's a minority opinion:
http://johnmacfarlane.net/babelmark2/?normalize=1&text=Text+text%0A-+one%0A-+two%0A

And, in fact, in our initial discussions I was pretty strongly
opposed to it.  Others felt strongly that people would naturally
create lists this way, and that it should be allowed. Some of
this is sumarized here:
http://talk.commonmark.org/t/requiring-a-blank-line-between-a-paragraph-and-list/657

The thread is a bit noisy but the point that sways me most is here:
http://talk.commonmark.org/t/requiring-a-blank-line-between-a-paragraph-and-list/657/13?u=jgm