Rename 'header' -> 'heading'.
This avoids a confusion that might arise now that HTML5 has
a 'header' element, distinct from the 'headings' h1, h2, ...
Our headings correspond to HTML5 headings, not HTML5 headers.

The terminology of 'headings' is more natural, too.

The only thing going for 'header' is that John Gruber used
it in his original Markdown syntax description.

See
http://talk.commonmark.org/t/naming-of-h1-6-should-be-headings-not-headers-per-w3c/1871