New format for spec tests, new lua formatter for specs.
The format for the spec examples has changed from

    .
    markdown
    .
    html
    .

to

    ```````````````````````````````` example
    markdown
    .
    html
    ````````````````````````````````

One advantage of this is that `spec.txt` becomes a valid
Markdown file.

`tests/spec_test.py` has been changed to use the new format.

The old `tools/makespec.py` has been replaced by a lua
program, `tools/make_spec.lua`, which uses the `lcmark` rock
(and indirectly libcmark).  It can generate
html, latex, and commonmark versions of the spec.
Pandoc is no longer needed for the latex/PDF version.
And, since the new program uses the cmark API and operates
directly on the parse tree, we avoid certain bad results we
got with the regex replacements done by the python script.