Spec changes in strong/emph parsing.
See
https://talk.commonmark.org/t/emphasis-strong-emphasis-corner-cases/2123
for motivation.

This restores intuitive parsings for a number of changes.

The main change is to disallow strong or emph when one of
the delimiters is "internal" and the sum of the lengths of
the enclosing delimiter runs is a multiple of 3.

Thus,

    **foo*bar***

gets parsed `<strong>foo*bar</strong>` rather than
`<em><em>foo</em>bar</em>**` as before.