Address the Gomphocarpus example
This example was not a correct use of strong or of em.
(It's a good showcase for why strong and em are wrong.)

Changed it to b and i, which the new rule 13 from the previous commit
now allows.

I'm putting this in a separate commit since another solution would be
to change it to foo and bar.