Add symbols to unicode punctuation
The current commonmark definition of “ASCII punctuation character” does not
only include the unicode punctuation characters within the ASCII range, but
also the unicode symbol characters within the ASCII range: `$`, `+`, `<`,
`=`, `>`, `^`, `` ` ``, `|`, `~`.

I think this makes sense. I think that users would not know the difference,
and would indeed expect to be able to escape `*` the same as `$`.

I think it is also sensical to broaden the definition of unicode punctuation
the same way, so that there’s no difference between `$`, `£`, and `€`.

Currently:

```markdown
*$*a.

*£*a.

*€*a.
```

*$*a.

*£*a.

*€*a.

Proposed: both do not turn into `<em>`s (note this is not a useful example,
I am sure it could be possible to come up with a better one).

> Note, `S` (symbol) includes the subgroups:
>
> ```
> gc ; Sc                               ; Currency_Symbol
> gc ; Sk                               ; Modifier_Symbol
> gc ; Sm                               ; Math_Symbol
> gc ; So                               ; Other_Symbol
> ```